Backend/requirements.txt
# FastAPI and web server
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Pydantic and settings
pydantic==2.5.0
pydantic-settings==2.1.0

# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.9 
alembic==1.12.1

# Environment and configuration
python-dotenv==1.0.0

# Data processing (Python 3.12 compatible versions)
pandas==2.2.0
numpy==1.26.4

# AI and ML (Phase 2) - Compatible versions
openai>=1.6.1
pinecone-client>=2.2.4
langchain>=0.0.335
langchain-openai>=0.0.2
langchain-pinecone>=0.0.1

# Caching (Phase 3)
redis>=5.0.1
hiredis>=2.2.3

# Utilities
python-json-logger==2.0.7
tenacity>=8.2.3
httpx>=0.25.2 